% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_analysis_utils.R
\name{calc_all_hmm_diff_parallel}
\alias{calc_all_hmm_diff_parallel}
\title{HMM segmentation for each gene}
\usage{
calc_all_hmm_diff_parallel(sorted_gene_exprs_matrix, cls_num = 2,
  threadnum = 1)
}
\arguments{
\item{sorted_gene_exprs_matrix}{Result returned from gene_selection function. A re-ordered gene by cell matrix}

\item{cls_num}{The number of states for HMM. Defaults to 2.}

\item{threadnum}{The number of threads to use.}
}
\description{
HMM is inferred for a ON/HIGH and OFF/LOW state for each significant gene along a developmental pseudo time. Extension to a 3-state model is also provided.
}
\section{Biological Analysis}{

}

\examples{
dcor_result <- calc_correlation(full_exprs_matrix, redpath_pseudotime, type = "dcor") 
mic_result <- calc_correlation(full_exprs_matrix, redpath_pseudotime, type = "mic")
selected_result <- gene_selection(full_exprs_matrix, redpath_pseudotime, dcor_result, mic_result, threshold = 0.5)
gene_state_result <- calc_all_hmm_diff_parallel(selected_result, cls_num = 2, threadnum = 4)
}
\keyword{gene_selection_intersection}
