% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_analysis_utils.R
\name{calc_correlation}
\alias{calc_correlation}
\title{Calculate correlation of gene expression with pseudotime}
\usage{
calc_correlation(full_gene_exprs, pseudo_time, type = "dcor",
  order = F)
}
\arguments{
\item{full_gene_exprs}{A gene (row) by cell (col) gene expression matrix with all genes.}

\item{pseudo_time}{The pseudo development time / ordering of cells can be provided as input. Please leave order = F if pseudotime is provided}

\item{type}{The type of correlation to calculate. "dcor" and "mic" are available (Distance correlation / Maximal Information Criterion).}

\item{order}{Defaults to F. If pseudotime is provided, set to F. If an ordering of cells is provided, set to T.}
}
\description{
This function calculates the correlation of gene expression with the inferred pseudo time.
}
\section{Biological Analysis}{

}

\examples{
dcor_result <- calc_correlation(full_exprs_matrix, redpath_pseudotime, type = "dcor") 
mic_result <- calc_correlation(full_exprs_matrix, redpath_pseudotime, type = "mic")
selected_result <- gene_selection(full_exprs_matrix, redpath_pseudotime, dcor_result, mic_result, threshold = 0.5)
}
\keyword{calc_correlation}
