% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{change_index}
\alias{change_index}
\title{Change Index}
\usage{
change_index(cell_labels, pseudotime, order = F)
}
\arguments{
\item{cell_labels}{The cell type labels}

\item{pseudotime}{The pseudo time / order index calculated from various softwares.}

\item{order}{Defaults to FALSE for input of pseudo-time, TRUE for input of an order index of cells}
}
\description{
It calculates the change index of the ordered cells. Evaluation in the range of [0, 1] with 1 being perfectly aligned.
}
\section{Evaluation}{

}

\examples{
ci <- change_index(cell_labels, redpath_pseudotime, order = F) 
}
\keyword{change_index}
