% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g0_identification.R
\name{distribution_plot}
\alias{distribution_plot}
\title{Visualization of G0-like distribution in mean scores}
\usage{
distribution_plot(mean_scores, kmean_result)
}
\arguments{
\item{mean_scores}{The mean score results returned from the reCAT software, "get_score"}

\item{kmean_result}{The classification of kmeans clustering result.}
}
\description{
Provides boxplot for visualizing the cell cluster distribution across different cell cycle stages mean scores
}
\section{G0 Identification}{

}

\examples{
cycle_mean_scores <- get_score(t(test_exp))$mean_score
kmean_classification <- kmeans_clustering(cycle_mean_scores)
require(scater)
distribution_plot(cycle_mean_scores, kmean_classification)
}
\keyword{distribution_plot}
