% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_visualization.R
\name{get_branch_viz_data}
\alias{get_branch_viz_data}
\title{Prepares the data for visualizing linear or branching process}
\usage{
get_branch_viz_data(all_hmt_soln, pseudotime, labels)
}
\arguments{
\item{all_hmt_soln}{The Hamiltonian path solutions of your data, stored as 'norm_res_new' after running the redpath() function.}

\item{pseudotime}{The inferred pseudo development time}

\item{labels}{The cell type / time-point labels.}
}
\description{
The function takes the matrix of multiple Hamiltonian path solutions, which is automatically saved to the global environment as 'norm_res_new' after running the redpath function.
It calculates the transition matrix and prepares the PCA coordinates for visualizing the linear / branching process.
}
\section{Biological Analysis}{

}

\examples{
redpath_pseudotime <- redpath(data, threadnum = 4) 
# Note that norm_res_new is stored as a global variable and it will be overwritten each time when you perform redpath. 
all_hamiltonian_solutions <- norm_res_new
branching_visualization_data <- get_branch_viz_data(all_hamiltonian_solutions, pseudotime = redpath_pseudotime, labels = NULL)
# Plot to visualize the trajectory
traj_plots <- plot_trajectory(branching_visualization_data, redpath_pseudotime, path = F, rev = F)
library(gridExtra)
grid.arrange(traj_plots[[1]], traj_plots[[2]], layout = matrix(c(1:2), 1, 2)) 
}
\keyword{get_branch_viz_data}
