% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_analysis_plots.R
\name{get_specific_gene_exprs}
\alias{get_specific_gene_exprs}
\title{Search for specific gene expression}
\usage{
get_specific_gene_exprs(full_gene_exprs_matrix, gene_name = "cdk",
  type = "grep", organism = "mm")
}
\arguments{
\item{full_gene_exprs_matrix}{This is the full gene (row) by cell (col) matrix of your dataset.}

\item{gene_name}{The gene name that is desired / enquired.}

\item{type}{pattern matching "grep" or exact "match"}

\item{organism}{"mm" for Mouse or "hs" for Human}
}
\description{
Gets the subset of genes which resemble the enquired gene name within the dataset. 
This prepares for the plotting function of multiple gene expressions along the pseudo development time.
}
\section{Biological Analysis}{

}

\examples{
cdk_gene_exprs <- get_specific_gene_exprs(full_gene_exprs_data, gene_name = "cdk11", type = "grep", organism = "mm")
plot_cdk <- plot_specific_gene_exprs(cdk_gene_exprs, redpath_pseudotime, color_label = cell_type_label, order = T)
require(scater)
multiplot(plotlist = plot_cdk, layout = matrix(c(1:2), 2, 1)) # For a total number of 2 plots
}
\keyword{get_specific_gene_exprs}
