% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_trial.R
\name{get_start_end}
\alias{get_start_end}
\title{Get start & end of path}
\usage{
get_start_end(dist_mat, k = 6)
}
\arguments{
\item{dist_mat}{A N by N distance matrix (KL distance in this case). Works on any other asymmetrical / symmetrical distance matrices}
}
\description{
This function estimates the initial starting and end node as input to the arbitrary insertion algorithm
}
\section{redPATH pseudotime}{

}

\examples{
se <- get_start_end(kl_cpp(data))
hamiltonian_path_ai(kl_cpp(data), se)

}
\keyword{get_start_end}
