% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_trial.R
\name{hamiltonian_path_ai}
\alias{hamiltonian_path_ai}
\title{Hamiltonian Path - Modified Arbitrary Insertion}
\usage{
hamiltonian_path_ai(dist_mat, se)
}
\arguments{
\item{dist_mat}{A N by N distance matrix (KL distance in this case). Works on any other asymmetrical / symmetrical distance matrices}
}
\description{
This function calculates the Hamiltonian path solution with a modified arbitrary insertion algorithm - designed for asymmetric cases (eg a KL distance matrix)
}
\section{redPATH pseudotime}{

}

\examples{
se <- get_start_end(kl_cpp(data))
hamiltonian_path_ai(kl_cpp(data), se)

}
\keyword{hamiltonian_path_as}
