% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_trial_new.R
\name{hamiltonian_path_as}
\alias{hamiltonian_path_as}
\title{Hamiltonian Path - Modified Arbitrary Insertion}
\usage{
hamiltonian_path_as(dist_mat, start = NULL)
}
\arguments{
\item{dist_mat}{A N by N distance matrix (KL distance in this case). Works on any other asymmetrical / symmetrical distance matrices}
}
\description{
This function calculates the Hamiltonian path solution with a modified arbitrary insertion algorithm - designed for asymmetric cases (eg a KL distance matrix)
}
\section{redPATH pseudotime}{

}

\examples{
hamiltonian_path_as(kl_cpp(data))

}
\keyword{hamiltonian_path_as}
