% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_cor_path.R
\name{hamiltonian_path_force}
\alias{hamiltonian_path_force}
\title{Hamiltonian path - by brute force}
\usage{
hamiltonian_path_force(distance_matrix)
}
\arguments{
\item{distance_matrix}{A N by N distance matrix (KL distance in this case). Works on any other asymmetrical / symmetrical distance matrices}
}
\description{
This function calculates the hamiltonian path by brute force. It's only scalable up to N = 10.
}
\section{redPATH pseudotime}{

}

\examples{
hamiltonian_path_as(kl_cpp(data))

}
\keyword{hamiltonian_path_force}
