% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_analysis_plots.R
\name{infer_GO_summaries}
\alias{infer_GO_summaries}
\title{Infer Gene Ontology summaries}
\usage{
infer_GO_summaries(gene_cluster, organism = "hsapiens")
}
\arguments{
\item{gene_cluster}{Gene cluster classification, with gene names.}

\item{organism}{Type of species of the genes. Default: "hsapiens"}
}
\description{
Relying on GOsummaries package, word cloud plots are generated using this function.
}
\section{Biological Analysis}{

}

\examples{
selected_result <- gene_selection(full_exprs_matrix, redpath_pseudotime, dcor_result, mic_result, threshold = 0.5)
gene_state_result <- calc_all_hmm_diff_parallel(selected_result, cls_num = 2, threadnum = 4) 
inferred_gene_cluster <- get_gene_cluster(sorted_matrix = gene_state_result, cls_num = 2)
GO_summary_result <- infer_GO_summaries(inferred_gene_cluster, organism = "hsapiens")
plot(GO_summary_result, fontsize = 8)
}
\keyword{infer_GO_summaries}
