% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{kendall_correlation}
\alias{kendall_correlation}
\title{Kendall Correlation}
\usage{
kendall_correlation(cell_labels_int, pseudotime, order = F)
}
\arguments{
\item{cell_labels_int}{The cell type labels written in terms of int, with ordering of cell types. For example, qNSC -> aNSC -> NPC becomese 1 -> 2 -> 3.}

\item{pseudotime}{The pseudo time / order index calculated from various softwares.}

\item{order}{Defaults to FALSE for input of pseudo-time, TRUE for input of an order index of cells}
}
\description{
It calculates the highest kendall correlation of the ordered cells. Evaluation in the range of [-1, 1] with 1 being perfectly aligned.
}
\section{Evaluation}{

}

\examples{
kc <- kendall_correlation(cell_labels_int, redpath_pseudotime, order = F) 

}
\keyword{kendall_correlation}
