% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_dist_optimized.R
\name{kl_cpp}
\alias{kl_cpp}
\title{Fast implementation of KL distance}
\usage{
kl_cpp(object, eps = 10^-4, overlap = TRUE, ...)
}
\arguments{
\item{object}{Input matrix}
}
\description{
This calculates the KL-distance between cells, generating the asymmetric distance matrix
}
\section{redPATH pseudotime}{

}

\examples{
kl_mat <- kl_cpp(data)
}
\keyword{kl_cpp}
