% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g0_identification.R
\name{kmeans_clustering}
\alias{kmeans_clustering}
\title{Kmeans Clustering - returns classification result}
\usage{
kmeans_clustering(mean_scores, k = 5, nstart = 100)
}
\arguments{
\item{mean_scores}{The mean score results returned from the reCAT software, "get_score"}

\item{k}{The number of clusters, defaults to 5. ie G0, G1, S, G2, M cycle stages}

\item{nstart}{Defaults to 100. Number of times to start the kmeans clustering, ensure stability of clustering results.}
}
\description{
This function performs imple kmeans clustering and prepares as input for identification of G0 cells
}
\section{G0 Identification}{

}

\examples{
cycle_mean_scores <- get_score(t(test_exp))$mean_score
kmean_classification <- kmeans_clustering(cycle_mean_scores)
statistical_test(cycle_mean_scores, kmean_classification, threshold = 0.001)
# Sample output
#"Possible group of G0-like cells is: 2"
}
\keyword{kmeans_clustering}
