% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{llorens_labels}
\alias{llorens_labels}
\title{Llorens-B - single cell neural stem cells dataset.}
\format{A vector of 145 cell type labels}
\usage{
llorens_labels
}
\description{
A dataset of 145 cells. 
 Referenced from:
 Llorens-Bobadilla, E., Zhao, S., Baser, A., Saiz-Castro, G., Zwadlo, K. and Martin-Villalba, A. (2015) 
 Single-Cell Transcriptomics Reveals a Population of Dormant Neural Stem Cells that Become Activated upon Brain Injury. Cell Stem Cell, 17, 329-340.
}
\keyword{datasets}
