% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g0_identification.R
\name{mean_score_plot}
\alias{mean_score_plot}
\title{Visualization of G0-like distribution in mean scores along a cell cycle time-series. 
6 plots for each of G1, S, G1/S, G2, M, G2/M mean scores.}
\usage{
mean_score_plot(mean_scores, kmean_result, ordIndex)
}
\arguments{
\item{mean_scores}{The mean score results returned from the reCAT software, "get_score"}

\item{kmean_result}{The classification of kmeans clustering result.}

\item{ordIndex}{The order obtained from reCAT, "get_ordIndex" which returns the cell cycle time-series of the data.}
}
\description{
Provides mean score expression changees along the cell cycle time-series with classification of different cell clusters.
}
\section{G0 Identification}{

}

\examples{
cycle_mean_scores <- get_score(t(test_exp))$mean_score
kmean_classification <- kmeans_clustering(cycle_mean_scores)
reCAT_ordIndex <- get_ordIndex(test_exp, threadnum = 2)
require(scater)
mean_score_plot(cycle_mean_scores, kmean_classification, reCAT_ordIndex)
}
\keyword{mean_score_plot}
