% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mgh107_portal_labels}
\alias{mgh107_portal_labels}
\title{MGH107 - single cell glioma dataset of MGH107 patient.}
\format{A vector of 252 cell type labels}
\usage{
mgh107_portal_labels
}
\description{
A dataset of 252 cells.
 Referenced from:
 Venteicher, A.S., Tirosh, I., Hebert, C., Yizhak, K., Neftel, C., Filbin, M.G., Hovestadt, V., Escalante, L.E., Shaw, M.L., Rodman, C. et al. (2017) 
 Decoupling genetics, lineages, and microenvironment in IDH-mutant gliomas by single-cell RNA-seq. Science, 355
}
\keyword{datasets}
