% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_analysis_plots.R
\name{plot_cycle_diff_3d}
\alias{plot_cycle_diff_3d}
\title{Coupling cell cycle and differentiation 3-D plots}
\usage{
plot_cycle_diff_3d(recat_input, redpath_pseudotime, recat_symbol,
  color_label)
}
\arguments{
\item{recat_input}{This is the normalizedResultLst obtained from reCAT software. It's generally save in a .RData file started with bestEnsemble...}

\item{redpath_pseudotime}{Pseudo development time for differentiation}

\item{recat_symbol}{Cell cycle labels}

\item{color_label}{Cell type labels}
}
\description{
This function incorporates output from reCAT and redPATH to illustrate some relationship between cell proliferation & differentiation.
}
\section{Biological Analysis}{

}

\examples{
p1 <- plot_cycle_diff_3d(normalizedResultLst, redpath_pseudotime, cell_cycle_labels, cell_type_labels)
p1
}
\keyword{plot_cycle_diff_3d}
