% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_analysis_plots.R
\name{plot_diff_3d}
\alias{plot_diff_3d}
\title{Coupling cell cycle and differentiation 3-D plots - Gene Exprs}
\usage{
plot_diff_3d(recat_input, redpath_pseudotime, color_label = NULL,
  marker_gene_exprs = NULL)
}
\arguments{
\item{recat_input}{This is the normalizedResultLst obtained from reCAT software. It's generally save in a .RData file started with bestEnsemble...}

\item{redpath_pseudotime}{Pseudo development time for differentiation}

\item{color_label}{Cell type labels}

\item{marker_gene_exprs}{Marker genes matrix (genes of interest) obtained from "get_specific_gene_exprs"}
}
\description{
This function incorporates output from reCAT and redPATH to illustrate some relationship between cell proliferation & differentiation. 
Observing the gradual change of specific marker genes
}
\section{Biological Analysis}{

}

\examples{
p1 <- plot_diff_3d(normalizedResultLst, redpath_pseudotime, cell_cycle_labels, cdk_gene_exprs)
p1
}
\keyword{plot_diff_3d}
