% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_analysis_plots.R
\name{plot_hmm_full}
\alias{plot_hmm_full}
\title{Plot function for HMM & gene expression changes}
\usage{
plot_hmm_full(sorted_exprs_val, pseudotime, sorted_hmm_results = NULL,
  color_label = NULL, num_plots = 10)
}
\arguments{
\item{sorted_exprs_val}{The gene expression input obtained from gene_selection. The ranked and sorted gene_expression matrix.}

\item{pseudotime}{Pseudo development time}

\item{sorted_hmm_results}{Defaults to NULL. This is the inferred HMM matrix obtained from calc_all_hmm_diff_parallel.}

\item{color_label}{Defaults to NULL. The cell type labels.}

\item{num_plots}{Number of plots desired, ranking from the most significant.}
}
\description{
Various versions of plots are available here to visualize the most significant gene expression change from dCor and MIC.
}
\section{Biological Analysis}{

}

\examples{
selected_result <- gene_selection(full_exprs_matrix, redpath_pseudotime, dcor_result, mic_result, threshold = 0.5)
gene_state_result <- calc_all_hmm_diff_parallel(selected_result, cls_num = 2, threadnum = 4)
hmm_plots <- plot_hmm_full(selected_result, redpath_pseudotime, gene_state_result, cell_type_label, num_plots = 10)
require(scater)
multiplot(plotlist = hmm_plots, layout = matrix(c(1:10), 5, 2)) # For a total number of 10 plots
}
\keyword{plot_hmm_full}
