% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_analysis_plots.R
\name{plot_specific_gene_exprs}
\alias{plot_specific_gene_exprs}
\title{Plot function for specific gene expression changes}
\usage{
plot_specific_gene_exprs(selected_gene, pseudotime, color_label,
  order = T)
}
\arguments{
\item{selected_gene}{The enquired specific gene expression matrix, obtained from "get_specific_gene_exprs".}

\item{pseudotime}{Pseudo development time}

\item{color_label}{The cell type labels.}

\item{order}{If order = T, gene expression change will be plotted against a uniform pseudo time. If order F, it will be plotted against the actual value of the calculated pseudo time.}
}
\description{
Plotting function for specific genes of interest, along the pseudo development time.
}
\section{Biological Analysis}{

}

\examples{
cdk_gene_exprs <- get_specific_gene_exprs(full_gene_exprs_data, gene_name = "cdk11", type = "grep", organism = "mm")
plot_cdk <- plot_specific_gene_exprs(cdk_gene_exprs, redpath_pseudotime, color_label = cell_type_label, order = T)
require(scater)
multiplot(plotlist = plot_cdk, layout = matrix(c(1:2), 2, 1)) # For a total number of 2 plots
}
\keyword{plot_specific_gene_exprs}
