% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_visualization.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plotting function for visualization of the trajectory}
\usage{
plot_trajectory(prepared_data, pseudotime, path = F, rev = F)
}
\arguments{
\item{prepared_data}{The data returned from the get_branch_viz_data() function.}

\item{pseudotime}{The inferred pseudo development time}

\item{path}{Defaults to F. Decides whether to apply the geom_path function on the plots.}

\item{rev}{Defaults to F. Sets the forward or reverse direction of the pseudotime.}
}
\description{
This function takes the prepared data from the 'get_branch_viz_data' function. 
It produces two PCA plots, showing the pseudotime information and cell type information separately.
}
\section{Biological Analysis}{

}

\examples{
redpath_pseudotime <- redpath(data, threadnum = 4) 
# Note that norm_res_new is stored as a global variable and it will be overwritten each time when you perform redpath. 
all_hamiltonian_solutions <- norm_res_new
branching_visualization_data <- get_branch_viz_data(all_hamiltonian_solutions, pseudotime = redpath_pseudotime, labels = NULL)
# Plot to visualize the trajectory
traj_plots <- plot_trajectory(branching_visualization_data, redpath_pseudotime, path = F, rev = F)

library(gridExtra)
grid.arrange(traj_plots[[1]], traj_plots[[2]], layout = matrix(c(1:2), 1, 2)) 
}
\keyword{plot_trajectory}
