% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{pos_score}
\alias{pos_score}
\title{POS Score}
\usage{
pos_score(cell_labels, pseudotime, order = F)
}
\arguments{
\item{pseudotime}{The pseudo time / order index calculated from various softwares.}

\item{order}{Defaults to FALSE for input of pseudo-time, TRUE for input of an order index of cells}

\item{cell_labels_int}{The cell type labels written in terms of int, with ordering of cell types. For example, qNSC -> aNSC -> NPC becomese 1 -> 2 -> 3.}
}
\description{
Adapted from Zhicheng Ji and Hongkai Ji. (2016)
TSCAN: Pseudo-time reconstruction and evaluation in single-cell RNA-seq analysis. (2016) Nucleic Acids Research, 44(13):e117.
Calculates the POS score for evaluating the pseudotime alignment. This is modified to give a value between [0, 1].
}
\section{Evaluation}{

}

\examples{
pos <- pos_score(cell_labels_int, redpath_pseudotime, order = F) 

}
\keyword{pos_score}
