% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redpath_pseudotime.R
\name{redpath}
\alias{redpath}
\title{redPATH Pseudotime}
\usage{
redpath(cell_gene_exprs, threadnum, step_size = 1,
  base_path_range = c(3:7), dist_type = "KL2", clustMethod = "Corr")
}
\arguments{
\item{cell_gene_exprs}{The gene expression input obtained from get_GO_exp}

\item{threadnum}{The number of threads to use}

\item{step_size}{The interval of paths to merge in the final solution. Defaults to 1.}

\item{base_path_range}{Sets the range of any consequtive 5 cluster solutions to calculate the base path. The maximum number can be 10 (ie c(6:10)). Defaults to c(3:7).}

\item{clustMethod}{= "Corr" This sets the clustering method used. Defaults to "Corr", which is hierarchical clustering on doubly centered matrix. "GMM" is also available for Gaussian Mixture Model.}
}
\description{
This function performs the redPATH algorithm for single cell gene expression data, obtaining the pseudo development time of each single cell.
}
\section{redPATH pseudotime}{

}

\examples{
redpath_pseudotime <- redpath(data, threadnum = 4) 
}
\keyword{redpath}
