% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g0_identification.R
\name{statistical_test}
\alias{statistical_test}
\title{Identification of G0-like cells}
\usage{
statistical_test(mean_scores, kmean_result, threshold = 0.001)
}
\arguments{
\item{mean_scores}{The mean score results returned from the reCAT software, "get_score"}

\item{kmean_result}{The classification of kmeans clustering result.}

\item{threshold}{Defaults to 0.001. The criterion that all p-values must be less than the threshold in order to determine G0-like cells}
}
\description{
This function performs the simple statistical tests to determine the group of G0 like cells within single cell datasets.
}
\section{G0 Identification}{

}

\examples{
cycle_mean_scores <- get_score(t(test_exp))$mean_score
kmean_classification <- kmeans_clustering(cycle_mean_scores)
statistical_test(cycle_mean_scores, kmean_classification, threshold = 0.001)
# Sample output
#"Possible group of G0-like cells is: 2"

}
\keyword{statistical_test}
